Migrate Alert and Suspicion to its own class
The EnemyManager is already big enough as it is, so I moved the Alert and Suspicion stuff into its own class to hide some of the bloat. Plus I think it'll be useful to provide more methods in here later on
