+++ Assets/Scripts/Enemy/AlertSuspicionLevel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c7088e00033d10e408c3d6bc99f8fde8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
