+++ Assets/Scripts/Enemy/EnemyManager.cs
@@ -7,31 +7,7 @@ public class EnemyManager : MonoBehaviour
     static EnemyManager singleton; //Lazy singleton pattern because lazy rn
     public static EnemyManager Singleton => singleton;
 
-    public enum AlertLevel
-    {
-        NONE,
-        LOW,
-        MEDIUM,
-        HIGH,
-        LOCKDOWN
-    }
-    public AlertLevel Alert { get; private set; }
-    public delegate void AlertLevelDelegate(AlertLevel alertLevel);
-    public AlertLevelDelegate OnAlertLevelChanged;
-
-    public enum SuspicionLevel
-    { //If the player gets spotted frequently but doesn't get caught the SuspicionLevel goes up
-        ///potentially raises AlertLevel
-        NONE,
-        SLIGHT,
-        SEARCH,
-        REINFORCE,
-        ALERT, //This raises the AlertLevel and drops suspicion to search
-    }
-    public SuspicionLevel Suspicion { get; private set; }
-    public delegate void OnSuspicionLevelChanged(SuspicionLevel suspicion);
-    public OnSuspicionLevelChanged SuspicionLevelChanged;
-
+    public AlertSuspicionLevel AlertLevel = new AlertSuspicionLevel();
 
     /// <summary>
     /// If a player has been spotted, this is true. Could potentially raise the alert level if spotted again
@@ -66,21 +42,11 @@ public class EnemyManager : MonoBehaviour
             Debug.Log("You have two EnemyManagers in the scene, disabling this one.");
             enabled = false; //Turn the script off and let somebody know
         }
-        OnAlertLevelChanged = new AlertLevelDelegate(LogAlertChanged);
-        SuspicionLevelChanged = new OnSuspicionLevelChanged(LogSupsicionLevel);
         InvestigateChanged = new OnInvestigateChanged(LogInvestigateState);
         OnEnemySpawned = new OnEnemySpawn(LogEnemySpawn);
     }
 
-    void LogAlertChanged(AlertLevel level)
-    {
-        Debug.Log("Alert level changed. New alert level is: " + level.ToString());
-    }
 
-    void LogSupsicionLevel(SuspicionLevel level)
-    {
-        Debug.Log("Suspicion level changed. New suspicion level is " + level.ToString());
-    }
 
     void LogInvestigateState(bool investigating)
     {
